requirements.txt
GitPython
regex
